+++ kwssdk/src/test/java/kws/superawesome/tv/kwssdk/TestSuite.java
@@ -18,18 +18,18 @@ import kws.superawesome.tv.kwssdk.requests.TestScoreRequest;
 import kws.superawesome.tv.kwssdk.requests.TestSetAppDataWrapperRequest;
 import kws.superawesome.tv.kwssdk.requests.TestTriggerEventRequest;
 import kws.superawesome.tv.kwssdk.requests.TestUserDetailsRequest;
-import kws.superawesome.tv.kwssdk.services.app.TestAppService_GetAppDataWrapper;
-import kws.superawesome.tv.kwssdk.services.app.TestAppService_GetLeadersWrapper;
-import kws.superawesome.tv.kwssdk.services.app.TestAppService_SetAppDataWrapper;
-import kws.superawesome.tv.kwssdk.services.create_user.TestAuthUserResponseService;
-import kws.superawesome.tv.kwssdk.services.events.TestEventsService_HasTriggeredEvent;
-import kws.superawesome.tv.kwssdk.services.events.TestEventsService_TriggerEvent;
-import kws.superawesome.tv.kwssdk.services.login.TestLoginService_LoginAuthResponse;
-import kws.superawesome.tv.kwssdk.services.random_username.TestRandomUsernameService;
-import kws.superawesome.tv.kwssdk.services.user.TestUserService_GetScore;
-import kws.superawesome.tv.kwssdk.services.user.TestUserService_GetUserDetails;
-import kws.superawesome.tv.kwssdk.services.user.TestUserService_InviteUser;
-import kws.superawesome.tv.kwssdk.services.user.TestUserService_RequestPermissions;
+import kws.superawesome.tv.kwssdk.services.odlServices.app.TestAppService_GetAppDataWrapper;
+import kws.superawesome.tv.kwssdk.services.odlServices.app.TestAppService_GetLeadersWrapper;
+import kws.superawesome.tv.kwssdk.services.odlServices.app.TestAppService_SetAppDataWrapper;
+import kws.superawesome.tv.kwssdk.services.odlServices.create_user.TestAuthUserResponseService;
+import kws.superawesome.tv.kwssdk.services.odlServices.events.TestEventsService_HasTriggeredEvent;
+import kws.superawesome.tv.kwssdk.services.odlServices.events.TestEventsService_TriggerEvent;
+import kws.superawesome.tv.kwssdk.services.odlServices.login.TestLoginService_LoginAuthResponse;
+import kws.superawesome.tv.kwssdk.services.odlServices.random_username.TestRandomUsernameService;
+import kws.superawesome.tv.kwssdk.services.odlServices.user.TestUserService_GetScore;
+import kws.superawesome.tv.kwssdk.services.odlServices.user.TestUserService_GetUserDetails;
+import kws.superawesome.tv.kwssdk.services.odlServices.user.TestUserService_InviteUser;
+import kws.superawesome.tv.kwssdk.services.odlServices.user.TestUserService_RequestPermissions;
 
 /**
  * Created by guilherme.mota on 09/01/2018.
