+++ kwssdk/src/test/java/kws/superawesome/tv/kwssdk/services/odlServices/app/TestAppService.java
@@ -1,23 +1,25 @@
-package kws.superawesome.tv.kwssdk.services.events;
+package kws.superawesome.tv.kwssdk.services.odlServices.app;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
 import kws.superawesome.tv.kwssdk.base.KWSSDK;
-import kws.superawesome.tv.kwssdk.base.services.EventsService;
+import kws.superawesome.tv.kwssdk.base.services.AppService;
 import kws.superawesome.tv.kwssdk.services.TestBaseService;
 
 /**
  * Created by guilherme.mota on 15/01/2018.
  */
 
-public class TestEventsService extends TestBaseService {
+public class TestAppService extends TestBaseService {
 
     //class to test
-    protected EventsService service;
+    protected AppService service;
 
-    //given
+    protected int goodAppId = 2;
+    protected int badAppId = 0;
+
     protected int goodUserId = 25;
     protected int badUserId = 0;
 
@@ -33,7 +35,8 @@ public class TestEventsService extends TestBaseService {
 
         //when
         // init class to test
-        service = KWSSDK.get(environment, EventsService.class, task);
+        service = KWSSDK.get(environment, AppService.class, task);
+
 
     }
