+++ kwssdk/src/test/java/kws/superawesome/tv/kwssdk/services/odlServices/events/TestEventsService_HasTriggeredEvent.java
@@ -1,4 +1,4 @@
-package kws.superawesome.tv.kwssdk.services.events;
+package kws.superawesome.tv.kwssdk.services.odlServices.events;
 
 import org.junit.Assert;
 import org.junit.Test;
